+++ grocery/src/main/java/com/example/grocery/core/security/services/UserDetailsImpl.java
@@ -16,7 +16,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private static final long serialVersionUID = 1L;
 
-    private int id;
+    private Long id;
     private String username;
     private String email;
     @JsonIgnore
@@ -28,7 +28,7 @@ public class UserDetailsImpl implements UserDetails {
         return authorities;
     }
 
-    public UserDetailsImpl(int id, String username, String email, String password,
+    public UserDetailsImpl(Long id, String username, String email, String password,
             Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
@@ -49,7 +49,7 @@ public class UserDetailsImpl implements UserDetails {
                 authorities);
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
