+++ grocery/src/main/java/com/example/grocery/dataAccess/abstracts/CorporateCustomerRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.grocery.entity.concretes.CorporateCustomer;
 
-public interface CorporateCustomerRepository extends JpaRepository<CorporateCustomer, Integer> {
+public interface CorporateCustomerRepository extends JpaRepository<CorporateCustomer, Long> {
 
     boolean existsByTaxNumber(String taxNumber);
 }
