+++ grocery/src/main/java/com/example/grocery/dataAccess/abstracts/ProducerRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.grocery.entity.concretes.Producer;
 
-public interface ProducerRepository extends JpaRepository<Producer, Integer> {
+public interface ProducerRepository extends JpaRepository<Producer, Long> {
 
     boolean existsByNameIgnoreCase(String name);
 }
