+++ grocery/src/main/java/com/example/grocery/dataAccess/abstracts/ProductRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.grocery.entity.concretes.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    boolean existsById(int id);
+    boolean existsById(Long id);
 
     boolean existsByNameIgnoreCase(String name);
 }
