+++ grocery/src/main/java/com/example/grocery/webApi/requests/order/CreateOrderRequest.java
@@ -28,12 +28,12 @@ public class CreateOrderRequest {
     private OrderStatus orderStatus;
 
     @Positive
-    private int paymentId;
+    private Long paymentId;
 
     @Positive
-    private int customerId;
+    private Long customerId;
 
     @NotNull
     // referans tipte @Positive olmaz alternatif bul...
-    private int[] productIds; // burayı Set ile yapılabilir...
+    private Long[] productIds; // burayı Set ile yapılabilir...
 }
