+++ grocery/src/main/java/com/example/grocery/webApi/requests/product/UpdateProductRequest.java
@@ -41,11 +41,11 @@ public class UpdateProductRequest {
     private int stock;
 
     @Positive
-    private int categoryId;
+    private Long categoryId;
 
     @Positive
-    private int supplierId;
+    private Long supplierId;
 
     @Positive
-    private int producerId;
+    private Long producerId;
 }
